Add refactored plot_issue_stats.py
The script was cleaned up and refactored heavily. It can now make use
of a local git repository to avoid querying Github for commits. The
list of issues is pickled locally and automatically updated to also
save network traffic.
Cleanup to possibly get rid of the numpy dependency is still WIP.